+++ cloudflare_tunnel/CHANGELOG.md
@@ -5,7 +5,35 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog][keepachangelog],
 and this project adheres to [Semantic Versioning][semver].
 
-## [Unreleased]
+## [1.6.0] - 2022-05-15
+
+### Added
+
+- Externalized ["origincert"][origincert]
+
+## [1.5.0] - 2022-05-14
+
+### Added
+
+- Apparmor
+
+## [1.4.1] - 2022-05-14
+
+### Changed
+
+- Disable apparmor temporarily
+
+## [1.4.0] - 2022-05-13
+
+### Added
+
+- Update for s6 v3 and add apparmor
+
+## [1.3.0] - 2022-02-13
+
+### Added
+
+- Added DNS routes automatically
 
 ## [1.1.0] - 2021-08-10
 
@@ -26,8 +54,6 @@ and this project adheres to [Semantic Versioning][semver].
 
 - Functionality for custom domains
 
-[unreleased]: https://github.com/w35l3y/hassio-addons/compare/v1.1.0...HEAD
-[1.1.0]: https://github.com/w35l3y/hassio-addons/compare/v1.0.1-1...v1.1.0
-[1.0.1]: https://github.com/w35l3y/hassio-addons/releases/tag/v1.0.1-1
 [semver]: https://semver.org/spec/v2.0.0.html
 [keepachangelog]: https://keepachangelog.com/en/1.0.0/
+[origincert]: https://developers.cloudflare.com/cloudflare-one/connections/connect-apps/configuration/arguments/#origincert